Move nightly test output code to nightly package
