+++ java/system-test/src/main/java/sleeper/systemtest/nightly/output/NightlyTestOutput.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.output;
+package sleeper.systemtest.nightly.output;
 
 import com.amazonaws.services.s3.AmazonS3;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
