+++ java/system-test/src/test/java/sleeper/systemtest/nightly/output/NightlyTestOutputS3IT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.output;
+package sleeper.systemtest.nightly.output;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
