+++ java/system-test/src/test/java/sleeper/systemtest/nightly/output/NightlyTestOutputTestHelper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.output;
+package sleeper.systemtest.nightly.output;
 
 import java.nio.file.Path;
 import java.util.List;
