+++ java/system-test/src/test/java/sleeper/systemtest/nightly/output/NightlyTestTimestampTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.output;
+package sleeper.systemtest.nightly.output;
 
 import org.junit.jupiter.api.Test;
