+++ scripts/test/runNightlyTests.sh
@@ -57,6 +57,6 @@ echo -n "$?" > "$OUTPUT_DIR/partitionSplitting.status"
 
 echo "[$(time_str)] Uploading test output"
 java -cp "${SYSTEM_TEST_JAR}" \
-sleeper.systemtest.output.RecordNightlyTestOutput "$RESULTS_BUCKET" "$START_TIMESTAMP" "$OUTPUT_DIR"
+sleeper.systemtest.nightly.output.RecordNightlyTestOutput "$RESULTS_BUCKET" "$START_TIMESTAMP" "$OUTPUT_DIR"
 
 popd
